+++ packages/format-switcher/src/Converter/KeyYamlToPhpFactory/ServicesKeyYamlToPhpFactory.php
@@ -6,7 +6,7 @@ namespace Migrify\ConfigTransformer\FormatSwitcher\Converter\KeyYamlToPhpFactory
 
 use Migrify\ConfigTransformer\FormatSwitcher\Contract\Converter\KeyYamlToPhpFactoryInterface;
 use Migrify\ConfigTransformer\FormatSwitcher\Contract\Converter\ServiceKeyYamlToPhpFactoryInterface;
-use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\ServicesPhpNodeFactory;
+use Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory\Service\ServicesPhpNodeFactory;
 use PhpParser\Node;
 
 /**
