--- packages/format-switcher/src/PhpParser/NodeFactory/NamespaceFactory.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Migrify\ConfigTransformer\FormatSwitcher\PhpParser\NodeFactory;
-
-final class NamespaceFactory
-{
-}
